+++ tests/testthat/test-bfs_get_data_comments.R
@@ -1,4 +1,7 @@
 test_that("bfs_get_data_comments() works", {
+  if (!curl::has_internet()) {
+    skip("No internet connection")
+  }
   df <- BFS::bfs_get_data_comments(
     number_bfs = "px-x-1502040100_131", 
     language = "en", 
@@ -7,6 +10,9 @@ test_that("bfs_get_data_comments() works", {
   expect_s3_class(df, "data.frame")
 })
 test_that("bfs_get_data_comments() query argument", {
+  if (!curl::has_internet()) {
+    skip("No internet connection")
+  }
   df <- BFS::bfs_get_data_comments(
     number_bfs = "px-x-1502040100_131", 
     language = "en", 
@@ -19,6 +25,9 @@ test_that("bfs_get_data_comments() query argument", {
   expect_s3_class(df, "data.frame")
 })
 test_that("bfs_get_data_comments() query argument not list", {
+  if (!curl::has_internet()) {
+    skip("No internet connection")
+  }
   expect_error(BFS::bfs_get_data_comments(
     number_bfs = "px-x-1502040100_131", 
     language = "en",